refactor: run forge fmt

+++ src/DSCEngine.sol
@@ -125,7 +125,6 @@ contract DSCEngine is ReentrancyGuard {
 
     function redeemCollateral() external {}
 
-
     /**
      * @notice Mints DSC to a user - Follow CEI
      * @param amountDSC The amount of DSC to mint
@@ -150,7 +149,11 @@ contract DSCEngine is ReentrancyGuard {
     }
 
     // Private & Internal Functions
-    function _getAccountInformation(address user) private view returns (uint256 totalDscMinted, uint256 collateralValueInUsd) {
+    function _getAccountInformation(address user)
+        private
+        view
+        returns (uint256 totalDscMinted, uint256 collateralValueInUsd)
+    {
         totalDscMinted = DSCMinted[user];
         collateralValueInUsd = getAccountCollateralValue(user);
     }
@@ -199,13 +202,13 @@ contract DSCEngine is ReentrancyGuard {
 
     function getUsdValue(address token, uint256 amount) public view returns (uint256) {
         AggregatorV3Interface priceFeed = AggregatorV3Interface(priceFeeds[token]);
-        (, int256 price ,,,) = priceFeed.latestRoundData();
+        (, int256 price,,,) = priceFeed.latestRoundData();
         return ((uint256(price) * ADDITIONAL_FEED_PRECISION) * amount) / PRECISION;
     }
 
     function getAccountCollateralValue(address user) public view returns (uint256) {
         uint256 totalCollateralValueInUsd;
-        for (uint i = 0; i < collateralTokens.length; i++) {
+        for (uint256 i = 0; i < collateralTokens.length; i++) {
             address token = collateralTokens[i];
             uint256 amount = collateralDeposited[user][token];
             totalCollateralValueInUsd += getUsdValue(token, amount);
